requirements.txt
Flask
Flask-Cors
PaddleOCR
Pillow
psycopg2
